Refactor tool implementations into their own modules
