+++ crates/notion-core/src/tool/mod.rs
@@ -1,6 +1,6 @@
 //! Traits and types for executing command-line tools.
 
-use std::env::{args_os, ArgsOs};
+use std::env::ArgsOs;
 use std::ffi::{OsStr, OsString};
 use std::fmt::{self, Debug, Display, Formatter};
 use std::io;
@@ -9,10 +9,23 @@ use std::path::Path;
 use std::process::{Command, ExitStatus};
 
 use notion_fail::{ExitCode, FailExt, Fallible, NotionError, NotionFail};
-use path;
 use session::{ActivityKind, Session};
 use style;
 use version::VersionSpec;
+
+mod binary;
+mod node;
+mod npm;
+mod npx;
+mod script;
+mod yarn;
+
+pub use self::binary::Binary;
+pub use self::node::Node;
+pub use self::npm::Npm;
+pub use self::npx::Npx;
+pub use self::script::Script;
+pub use self::yarn::Yarn;
 
 fn display_error(err: &NotionError) {
     if err.is_user_friendly() {
@@ -85,17 +98,6 @@ impl BinaryExecError {
     }
 }
 
-#[derive(Debug, Fail, NotionFail)]
-#[fail(display = "this tool is not yet implemented")]
-#[notion_fail(code = "ExecutableNotFound")]
-pub(crate) struct ToolUnimplementedError;
-
-impl ToolUnimplementedError {
-    pub(crate) fn new() -> Self {
-        ToolUnimplementedError
-    }
-}
-
 /// Represents a command-line tool that Notion shims delegate to.
 pub trait Tool: Sized {
     fn launch() -> ! {
@@ -153,157 +155,11 @@ pub trait Tool: Sized {
     }
 }
 
-/// Represents a delegated script.
-pub struct Script(Command);
-
-/// Represents a delegated binary executable.
-pub struct Binary(Command);
-
-/// Represents a Node executable.
-pub struct Node(Command);
-
-/// Represents a `npm` executable.
-pub struct Npm(Command);
-
-/// Represents a `npx` executable.
-pub struct Npx(Command);
-
-/// Represents a Yarn executable.
-pub struct Yarn(Command);
-
-#[cfg(windows)]
-impl Tool for Script {
-    fn new(_session: &mut Session) -> Fallible<Self> {
-        throw!(ToolUnimplementedError::new())
-    }
-
-    fn from_components(exe: &OsStr, args: ArgsOs, path_var: &OsStr) -> Self {
-        // The best way to launch a script in Windows is to use `cmd.exe`
-        // as the executable and pass `"/C"` followed by the name of the
-        // script and then its arguments. Unfortunately, the docs aren't
-        // super clear about this, but see the discussion at:
-        //
-        //     https://github.com/rust-lang/rust/issues/42791
-        let mut command = Command::new("cmd.exe");
-        command.arg("/C");
-        command.arg(exe);
-        command.args(args);
-        command.env("PATH", path_var);
-        Script(command)
-    }
-
-    fn command(self) -> Command {
-        self.0
-    }
-}
-
 fn command_for(exe: &OsStr, args: ArgsOs, path_var: &OsStr) -> Command {
     let mut command = Command::new(exe);
     command.args(args);
     command.env("PATH", path_var);
     command
-}
-
-#[cfg(unix)]
-impl Tool for Script {
-    fn new(_session: &mut Session) -> Fallible<Self> {
-        throw!(ToolUnimplementedError::new())
-    }
-
-    fn from_components(exe: &OsStr, args: ArgsOs, path_var: &OsStr) -> Self {
-        Script(command_for(exe, args, path_var))
-    }
-
-    fn command(self) -> Command {
-        self.0
-    }
-}
-
-#[derive(Debug, Fail, NotionFail)]
-#[fail(display = "No toolchain available to run shim {}", shim_name)]
-#[notion_fail(code = "ExecutionFailure")]
-pub(crate) struct NoToolChainError {
-    shim_name: String,
-}
-
-impl NoToolChainError {
-    pub(crate) fn for_shim(shim_name: String) -> NoToolChainError {
-        NoToolChainError { shim_name }
-    }
-}
-
-impl Tool for Binary {
-    fn new(session: &mut Session) -> Fallible<Self> {
-        session.add_event_start(ActivityKind::Binary);
-
-        let mut args = args_os();
-        let exe = arg0(&mut args)?;
-
-        // first try to use the project toolchain
-        if let Some(project) = session.project()? {
-            // check if the executable is a direct dependency
-            if project.has_direct_bin(&exe)? {
-                // use the full path to the file
-                let mut path_to_bin = project.local_bin_dir();
-                path_to_bin.push(&exe);
-
-                // if we're in a pinned project, use the project's platform.
-                if let Some(ref platform) = session.project_platform()? {
-                    let image = platform.checkout(session)?;
-                    return Ok(Self::from_components(
-                        &path_to_bin.as_os_str(),
-                        args,
-                        &image.path()?,
-                    ));
-                }
-
-                // otherwise use the user platform.
-                if let Some(ref platform) = session.user_platform()? {
-                    let image = platform.checkout(session)?;
-                    return Ok(Self::from_components(
-                        &path_to_bin.as_os_str(),
-                        args,
-                        &image.path()?,
-                    ));
-                }
-
-                // if there's no user platform selected, fail.
-                throw!(NoSuchToolError {
-                    tool: "Node".to_string()
-                });
-            }
-        }
-
-        // next try to use the user toolchain
-        if let Some(ref platform) = session.user_platform()? {
-            // use the full path to the binary
-            // ISSUE (#160): Look up the platform image bound to the user tool.
-            let image = platform.checkout(session)?;
-            let node_str = image.node.runtime.to_string();
-            let npm_str = image.node.npm.to_string();
-            let mut third_p_bin_dir = path::node_image_3p_bin_dir(&node_str, &npm_str)?;
-            third_p_bin_dir.push(&exe);
-            return Ok(Self::from_components(
-                &third_p_bin_dir.as_os_str(),
-                args,
-                &image.path()?,
-            ));
-        };
-
-        // at this point, there is no project or user toolchain
-        // the user is executing a Notion shim that doesn't have a way to execute it
-        throw!(NoToolChainError::for_shim(
-            exe.to_string_lossy().to_string()
-        ));
-    }
-
-    fn from_components(exe: &OsStr, args: ArgsOs, path_var: &OsStr) -> Self {
-        Binary(command_for(exe, args, path_var))
-    }
-
-    fn command(self) -> Command {
-        self.0
-    }
 }
 
 #[derive(Fail, Debug)]
@@ -336,155 +192,4 @@ See `notion help install` for help adding {} to your personal toolchain."#,
 #[notion_fail(code = "NoVersionMatch")]
 struct NoSuchToolError {
     tool: String,
-}
-
-impl Tool for Node {
-    fn new(session: &mut Session) -> Fallible<Self> {
-        session.add_event_start(ActivityKind::Node);
-
-        let mut args = args_os();
-        let exe = arg0(&mut args)?;
-        if let Some(ref platform) = session.current_platform()? {
-            let image = platform.checkout(session)?;
-            Ok(Self::from_components(&exe, args, &image.path()?))
-        } else {
-            throw!(NoSuchToolError {
-                tool: "Node".to_string()
-            });
-        }
-    }
-
-    fn from_components(exe: &OsStr, args: ArgsOs, path_var: &OsStr) -> Self {
-        Node(command_for(exe, args, path_var))
-    }
-
-    fn command(self) -> Command {
-        self.0
-    }
-}
-
-impl Tool for Yarn {
-    fn new(session: &mut Session) -> Fallible<Self> {
-        session.add_event_start(ActivityKind::Yarn);
-
-        let mut args = args_os();
-        let exe = arg0(&mut args)?;
-        if let Some(ref platform) = session.current_platform()? {
-            let image = platform.checkout(session)?;
-            Ok(Self::from_components(&exe, args, &image.path()?))
-        } else {
-            throw!(NoSuchToolError {
-                tool: "Yarn".to_string()
-            });
-        }
-    }
-
-    fn from_components(exe: &OsStr, args: ArgsOs, path_var: &OsStr) -> Self {
-        Yarn(command_for(exe, args, path_var))
-    }
-
-    fn command(self) -> Command {
-        self.0
-    }
-
-    /// Perform any tasks which must be run after the tool runs but before exiting.
-    fn finalize(session: &Session, maybe_status: &io::Result<ExitStatus>) {
-        if let Ok(_) = maybe_status {
-            if let Ok(Some(project)) = session.project() {
-                let errors = project.autoshim();
-
-                for error in errors {
-                    display_error(&error);
-                }
-            }
-        }
-    }
-}
-
-impl Tool for Npm {
-    fn new(session: &mut Session) -> Fallible<Self> {
-        session.add_event_start(ActivityKind::Npm);
-
-        let mut args = args_os();
-        let exe = arg0(&mut args)?;
-        if let Some(ref platform) = session.current_platform()? {
-            let image = platform.checkout(session)?;
-            Ok(Self::from_components(&exe, args, &image.path()?))
-        } else {
-            // Using 'Node' as the tool name since the npm version is derived from the Node version
-            // This way the error message will prompt the user to add 'Node' to their toolchain, instead of 'npm'
-            throw!(NoSuchToolError {
-                tool: "Node".to_string()
-            });
-        }
-    }
-
-    fn from_components(exe: &OsStr, args: ArgsOs, path_var: &OsStr) -> Self {
-        Npm(command_for(exe, args, path_var))
-    }
-
-    fn command(self) -> Command {
-        self.0
-    }
-
-    fn finalize(session: &Session, maybe_status: &io::Result<ExitStatus>) {
-        if let Ok(_) = maybe_status {
-            if let Ok(Some(project)) = session.project() {
-                let errors = project.autoshim();
-
-                for error in errors {
-                    display_error(&error);
-                }
-            }
-        }
-    }
-}
-
-#[derive(Debug, Fail, NotionFail)]
-#[fail(
-    display = r#"
-'npx' is only available with npm >= 5.2.0
-
-This project is configured to use version {} of npm."#,
-    version
-)]
-#[notion_fail(code = "ExecutableNotFound")]
-struct NpxNotAvailableError {
-    version: String,
-}
-
-impl Tool for Npx {
-    fn new(session: &mut Session) -> Fallible<Self> {
-        session.add_event_start(ActivityKind::Npx);
-
-        let mut args = args_os();
-        let exe = arg0(&mut args)?;
-        if let Some(ref platform) = session.current_platform()? {
-            let image = platform.checkout(session)?;
-
-            // npx was only included with Node >= 8.2.0. If less than that, we should include a helpful error message
-            let required_node = VersionSpec::parse_requirements(">= 5.2.0")?;
-            if required_node.matches(&image.node.npm) {
-                Ok(Self::from_components(&exe, args, &image.path()?))
-            } else {
-                throw!(NpxNotAvailableError {
-                    version: image.node.npm.to_string()
-                });
-            }
-        } else {
-            // Using 'Node' as the tool name since the npx version is derived from the Node version
-            // This way the error message will prompt the user to add 'Node' to their toolchain, instead of 'npx'
-            throw!(NoSuchToolError {
-                tool: "Node".to_string()
-            });
-        }
-    }
-
-    fn from_components(exe: &OsStr, args: ArgsOs, path_var: &OsStr) -> Self {
-        Npx(command_for(exe, args, path_var))
-    }
-
-    fn command(self) -> Command {
-        self.0
-    }
 }
